Use separate module for status
