+++ icinga2api/client.py
@@ -10,10 +10,10 @@ from __future__ import print_function
 import logging
 
 import icinga2api
-from icinga2api.base import Base
 from icinga2api.objects import Objects
 from icinga2api.actions import Actions
 from icinga2api.events import Events
+from icinga2api.status import Status
 from icinga2api.configfile import ClientConfigFile
 from icinga2api.exceptions import Icinga2ApiException
 
@@ -68,33 +68,3 @@ class Client(object):
             raise Icinga2ApiException(
                 'Neither username/password nor certificate defined.'
             )
-
-
-class Status(Base):
-    '''
-    Icinga 2 API status class
-    '''
-
-    base_url_path = 'v1/status'
-
-    def list(self, component=None):
-        '''
-        retrieve status information and statistics for Icinga 2
-
-        example 1:
-        list()
-
-        example 2:
-        list('IcingaApplication')
-
-        :param component: only list the status of this component
-        :type component: string
-        :returns: status information
-        :rtype: dictionary
-        '''
-
-        url = self.base_url_path
-        if component:
-            url += "/{}".format(component)
-
-        return self._request('GET', url)
